backend-python/requirements.txt
flask==2.3.3
minio==7.1.15
torch==2.1.0
transformers==4.33.0
accelerate==0.23.0
sentencepiece==0.1.99
protobuf==4.24.3
pdfplumber==0.10.2
python-docx==0.8.11
python-magic==0.4.27